chore: join room hook
